requirements.txt
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
supabase==2.4.2
httpx[http2]==0.26.0
supafunc==0.4.5
h2==4.1.0
openai==1.12.0
celery==5.3.4
redis==5.0.1 